feat: refactor table components to use shared configuration and styles
