+++ src/components/atoms/Table/TableEmptyState.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { TableCell, TableRow } from "@/components/ui/table";
-import { historyTableConfig } from "../../organisms/HistoryTable/HistoryTable.styles";
+import { historyTableConfig } from "@/shared/config/historyTableConfig";
 
 interface TableEmptyStateProps {
   message: string;
@@ -8,7 +8,7 @@ interface TableEmptyStateProps {
 
 export const TableEmptyState: React.FC<TableEmptyStateProps> = ({ message }) => (
   <TableRow>
-    <TableCell colSpan={historyTableConfig.length} className="text-center py-8 text-muted-foreground">
+    <TableCell colSpan={historyTableConfig?.length} className="text-center py-8 text-muted-foreground">
       {message}
     </TableCell>
   </TableRow>
