+++ src/components/atoms/Table/TableLoadingState.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { TableCell, TableRow } from "@/components/ui/table";
-import { historyTableConfig } from "../../organisms/HistoryTable/HistoryTable.styles";
+import { historyTableConfig } from "@/shared/config/historyTableConfig";
 
 export const TableLoadingState: React.FC = () => (
   <TableRow>
-    <TableCell colSpan={historyTableConfig.length} className="text-center py-8 text-muted-foreground">
+    <TableCell colSpan={historyTableConfig?.length} className="text-center py-8 text-muted-foreground">
       Cargando...
     </TableCell>
   </TableRow>
